backend/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
autopep8==2.0.2
boto3==1.26.139
botocore==1.29.139
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
cloudinary==1.32.0
cryptography==40.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.3.0
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
drf-spectacular==0.26.1
ecdsa==0.18.0
firebase-admin==6.2.0
frozenlist==1.3.3
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.54.0
grpcio-status==1.54.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.17.3
Markdown==3.4.3
msgpack==1.0.5
multidict==6.0.4
phonenumbers==8.13.15
proto-plus==1.22.2
protobuf==4.22.3
psycopg2==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.4.0
wrapt==1.15.0
yarl==1.8.2
